+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.6.22] - 2021-04-23
 ### Added
+- A new grade, MAX, which is above SSS. Has a 30% stat bonus. Will this stop the questions about the max grade? Probably not...
 - Config to disable WIP tooltip on some items
 
 ## [2.6.21] - 2021-04-21